requirements.txt
django-2.1.3
djangorestframework-3.9.0